Fixed printf format errors.  Assume that uid_t's are representable as unsigned longs.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@40492 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.bin/chpass/pw_yp.c
@@ -35,7 +35,7 @@
  * Center for Telecommunications Research
  * Columbia University, New York City
  *
- *	$Id: pw_yp.c,v 1.11 1997/07/29 15:45:36 wpaul Exp $
+ *	$Id: pw_yp.c,v 1.12 1997/12/18 15:27:43 bde Exp $
  */
 
 #ifdef YP
@@ -254,7 +254,7 @@ int use_yp (user, uid, which)
 	char ubuf[UT_NAMESIZE + 2];
 
 	if (which) {
-		snprintf(ubuf, sizeof(ubuf), "%lu", uid);
+		snprintf(ubuf, sizeof(ubuf), "%lu", (unsigned long)uid);
 		user = (char *)&ubuf;
 	}
